Extract `removePagePathTail`
