+++ packages/next/server/page-path-utils.ts
@@ -19,6 +19,24 @@ export function normalizePathSep(path: string): string {
 }
 
 /**
+ * Removes the file extension for a page and the trailing `index` if it exists
+ * making sure to not return an empty string. The page head is not touched
+ * and returned as it is passed. Examples:
+ *   - `/foo/bar/baz/index.js` -> `/foo/bar/baz`
+ *   - `/foo/bar/baz.js` -> `/foo/bar/baz`
+ *
+ * @param pagePath A page to a page file (absolute or relative)
+ * @param extensions Extensions allowed for the page.
+ */
+export function removePagePathTail(pagePath: string, extensions: string[]) {
+  return (
+    normalizePathSep(pagePath)
+      .replace(new RegExp(`\\.+(?:${extensions.join('|')})$`), '')
+      .replace(/\/index$/, '') || '/'
+  )
+}
+
+/**
  * Takes a page and transforms it into its file counterpart ensuring that the
  * output is normalized. Note this function is not idempotent because a page
  * `/index` can be referencing `/index/index.js` and `/index/index` could be
