+++ test/integration/production-swcminify/test/index.test.js
@@ -80,7 +80,7 @@ describe.skip('Production Usage with swcMinify', () => {
     ).toBe(true)
     expect(
       serverTrace.files.some((file) =>
-        file.includes('next/dist/server/normalize-page-path.js')
+        file.includes('next/dist/server/page-path-utils.js')
       )
     ).toBe(true)
     expect(
